# src/infrastructure/configs/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "Parking API",
    "description": "Implementation of Swagger with TypeScript"
  },
  "servers": [
    {
      "url": "http://localhost:2400",
      "description": ""
    }
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/healthcheck": {
      "get": {
        "tags": [
          "Healthcheck"
        ],
        "summary": "runs healthcheck for api",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/parking/{plate}": {
      "get": {
        "tags": [
          "Parking"
        ],
        "summary": "List parking entrances by plate",
        "description": "This endpoint will list all entrance registers in the parking lot",
        "parameters": [
          {
            "name": "plate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit per page",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return data for parking",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/parkingListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/parking": {
      "post": {
        "tags": [
          "Parking"
        ],
        "summary": "Create a entrance",
        "description": "This endpoint will create a entrance in the parking lot",
        "responses": {
          "200": {
            "description": "Return the initial data for parking",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/parkingCreateResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/parkingCreateRequest"
              }
            }
          }
        }
      }
    },
    "/parking/{id}/out": {
      "put": {
        "tags": [
          "Parking"
        ],
        "summary": "Left parking lot",
        "description": "This endpoint will register the exit in the parking lot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the data for parking",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/parkingInfoResponse"
                }
              }
            }
          },
          "402": {
            "description": "Parking fee has not been paid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Parking record not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/parking/{id}/pay": {
      "put": {
        "tags": [
          "Parking"
        ],
        "summary": "Left parking lot",
        "description": "This endpoint will pay the parking lot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the data for parking",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/parkingInfoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Parking fee has already been paid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Parking record not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "parkingCreateRequest": {
        "type": "object",
        "properties": {
          "plate": {
            "type": "string",
            "example": "string"
          }
        }
      },
      "parkingCreateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "string"
          },
          "plate": {
            "type": "string",
            "example": "string"
          },
          "entranceDate": {
            "type": "string",
            "example": "date"
          }
        }
      },
      "parkingInfoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "string"
          },
          "plate": {
            "type": "string",
            "example": "string"
          },
          "time": {
            "type": "string",
            "example": "string"
          },
          "paid": {
            "type": "string",
            "example": "boolean"
          },
          "left": {
            "type": "string",
            "example": "boolean"
          }
        }
      },
      "parkingListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "string"
                },
                "plate": {
                  "type": "string",
                  "example": "string"
                },
                "time": {
                  "type": "string",
                  "example": "string"
                },
                "paid": {
                  "type": "string",
                  "example": "boolean"
                },
                "left": {
                  "type": "string",
                  "example": "boolean"
                }
              }
            }
          }
        }
      },
      "errorResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "string"
          },
          "message": {
            "type": "string",
            "example": "string"
          }
        }
      }
    }
  }
}